Fix syntax error

+++ cogs/gods/__init__.py
@@ -76,14 +76,11 @@ class Gods(commands.Cog):
             if len(loot_ids) > 0:
                 await conn.execute(
                     'DELETE FROM loot WHERE id ANY($1) AND "user"=$2;',
-                    loot_ids
-                    ctx.author.id
+                    loot_ids,
+                    ctx.author.id,
                 )
             else:
-                await conn.execute(
-                    'DELETE FROM loot WHERE "user"=$1;',
-                    ctx.author.id
-                )
+                await conn.execute('DELETE FROM loot WHERE "user"=$1;', ctx.author.id)
             await conn.execute(
                 'UPDATE profile SET "favor"="favor"+$1 WHERE "user"=$2;',
                 value,
